re-do with output data

+++ rebuildlocator/esri-rebuild-locators.py
@@ -197,8 +197,11 @@ def publishLocator(info):
 
     print "Starting to publish the geocode service " + service_name  + "..."
 
-    os.remove(out_sddraft)
-    os.remove(out_service_definition)
+    if os.path.isfile( out_sddraft):
+        os.remove(out_sddraft)
+
+    if os.path.isfile( out_service_definition ):
+        os.remove(out_service_definition)
 
     #stagging
     out_service_definition = info['out_service_definition']
